Fixed overflow in the calculation of the number of inodes per group for filesystems with almost the maximum number of sectors.  The maxiumum is 2^31, but overflow is common for that size, and overflow normally occurred here at size (2^31 - 4096).

+++ sbin/newfs/mkfs.c
@@ -1059,7 +1059,7 @@ calcipg(cpg, bpcg, usedbp)
 	 * Prepare to scale by fssize / (number of sectors in cylinder groups).
 	 * Note that fssize is still in sectors, not filesystem blocks.
 	 */
-	ncyl = howmany(fssize, secpercyl);
+	ncyl = howmany(fssize, (u_int)secpercyl);
 	ncg = howmany(ncyl, cpg);
 	/*
 	 * Iterate a few times to allow for ipg depending on itself.
